requirements.txt
PySide6>=6.6
numpy>=1.25
scipy>=1.11
matplotlib>=3.8
paramiko>=3.3
PyYAML>=6.0
